third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-active-state-expected.txt

Gestures go here

Tests that tap gesture events set and clear the active state of elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

See https://bugs.webkit.org/show_bug.cgi?id=96060 for details

PASS successfullyParsed is true

TEST COMPLETE

Verify active isn't initially set
PASS getHoverActiveState(box) is "default"
Verify showPress, tap sets and clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
Verify showPress, tapCancel on the element sets and clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
Verify tap elsewhere still clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "default"
Verify tapCancel elsewhere still clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "default"
Verify that touchStart doesn't trigger active state
PASS getHoverActiveState(box) is "default"
Verify that touchStart/touchEnd doesn't cancel active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
PASS successfullyParsed is true

TEST COMPLETE
